requirements.txt
pytest
requests
pyyaml
pydantic
loguru